// isa-mrs-2022/Angular-Frontend/angular-frontend/src/app/models/report/report.ts
export class ReportDTO {

    numberCottageReservation: number;

    numberBoatReservation: number;

    numberAdventureReservation: number;

    priceCottageReservation: number;

    priceBoatReservation: number;

    priceAdventureReservation: number;

    sum: number;

    appProfit: number;

    finishSum: number;

}